Pylint fixes
